+++ src/editor/core/command/Command.ts
@@ -9,6 +9,8 @@ export class Command {
   public executeSelectAll: CommandAdapt['selectAll']
   public executeBackspace: CommandAdapt['backspace']
   public executeSetRange: CommandAdapt['setRange']
+  public executeReplaceRange: CommandAdapt['replaceRange']
+  public executeSetPositionContext: CommandAdapt['setPositionContext']
   public executeForceUpdate: CommandAdapt['forceUpdate']
   public executeBlur: CommandAdapt['blur']
   public executeUndo: CommandAdapt['undo']
@@ -83,8 +85,10 @@ export class Command {
   public executeSetGroup: CommandAdapt['setGroup']
   public executeDeleteGroup: CommandAdapt['deleteGroup']
   public executeLocationGroup: CommandAdapt['locationGroup']
+  public executeSetZone: CommandAdapt['setZone']
   public executeSetControlValue: CommandAdapt['setControlValue']
   public executeSetControlExtension: CommandAdapt['setControlExtension']
+  public executeSetControlHighlight: CommandAdapt['setControlHighlight']
   public getCatalog: CommandAdapt['getCatalog']
   public getImage: CommandAdapt['getImage']
   public getOptions: CommandAdapt['getOptions']
@@ -96,6 +100,7 @@ export class Command {
   public getRangeContext: CommandAdapt['getRangeContext']
   public getRangeRow: CommandAdapt['getRangeRow']
   public getRangeParagraph: CommandAdapt['getRangeParagraph']
+  public getKeywordRangeList: CommandAdapt['getKeywordRangeList']
   public getPaperMargin: CommandAdapt['getPaperMargin']
   public getSearchNavigateInfo: CommandAdapt['getSearchNavigateInfo']
   public getLocale: CommandAdapt['getLocale']
@@ -112,6 +117,8 @@ export class Command {
     this.executeSelectAll = adapt.selectAll.bind(adapt)
     this.executeBackspace = adapt.backspace.bind(adapt)
     this.executeSetRange = adapt.setRange.bind(adapt)
+    this.executeReplaceRange = adapt.replaceRange.bind(adapt)
+    this.executeSetPositionContext = adapt.setPositionContext.bind(adapt)
     this.executeForceUpdate = adapt.forceUpdate.bind(adapt)
     this.executeBlur = adapt.blur.bind(adapt)
     // 撤销、重做、格式刷、清除格式
@@ -193,6 +200,7 @@ export class Command {
     this.executeSetGroup = adapt.setGroup.bind(adapt)
     this.executeDeleteGroup = adapt.deleteGroup.bind(adapt)
     this.executeLocationGroup = adapt.locationGroup.bind(adapt)
+    this.executeSetZone = adapt.setZone.bind(adapt)
     // 获取
     this.getImage = adapt.getImage.bind(adapt)
     this.getOptions = adapt.getOptions.bind(adapt)
@@ -204,6 +212,7 @@ export class Command {
     this.getRangeContext = adapt.getRangeContext.bind(adapt)
     this.getRangeRow = adapt.getRangeRow.bind(adapt)
     this.getRangeParagraph = adapt.getRangeParagraph.bind(adapt)
+    this.getKeywordRangeList = adapt.getKeywordRangeList.bind(adapt)
     this.getCatalog = adapt.getCatalog.bind(adapt)
     this.getPaperMargin = adapt.getPaperMargin.bind(adapt)
     this.getSearchNavigateInfo = adapt.getSearchNavigateInfo.bind(adapt)
@@ -213,6 +222,7 @@ export class Command {
     // 控件
     this.executeSetControlValue = adapt.setControlValue.bind(adapt)
     this.executeSetControlExtension = adapt.setControlExtension.bind(adapt)
+    this.executeSetControlHighlight = adapt.setControlHighlight.bind(adapt)
     this.getControlValue = adapt.getControlValue.bind(adapt)
   }
 }