+++ src/editor/dataset/constant/Element.ts
@@ -61,6 +61,7 @@ export const EDITOR_ELEMENT_ZIP_ATTR: Array<keyof IElement> = [
   'listStyle',
   'listWrap',
   'groupIds',
+  'conceptId'
 ]
 
 export const TABLE_TD_ZIP_ATTR: Array<keyof ITd> = [