+++ src/editor/interface/Position.ts
@@ -28,6 +28,7 @@ export interface ICurrentPosition {
 export interface IGetPositionByXYPayload {
   x: number
   y: number
+  pageNo?: number
   isTable?: boolean
   td?: ITd
   tablePosition?: IElementPosition
@@ -46,6 +47,11 @@ export interface IPositionContext {
   tdId?: string
   trId?: string
   tableId?: string
+}
+
+export interface IComputeRowPositionPayload {
+  row: IRow
+  innerWidth: number
 }
 
 export interface IComputePageRowPositionPayload {