+++ src/main.ts
@@ -17,7 +17,8 @@ import Editor, {
   PageMode,
   PaperDirection,
   RowFlex,
-  TitleLevel
+  TitleLevel,
+  splitText
 } from './editor'
 import { Dialog } from './components/dialog/Dialog'
 import { formatPrismToken } from './utils/prism'
@@ -423,7 +424,8 @@ window.onload = function () {
           label: '文本',
           name: 'name',
           required: true,
-          placeholder: '请输入文本'
+          placeholder: '请输入文本',
+          value: instance.command.getRangeText()
         },
         {
           type: 'text',
@@ -442,7 +444,7 @@ window.onload = function () {
           type: ElementType.HYPERLINK,
           value: '',
           url,
-          valueList: name.split('').map(n => ({
+          valueList: splitText(name).map(n => ({
             value: n,
             size: 16
           }))
@@ -563,7 +565,7 @@ window.onload = function () {
         const elementList: IElement[] = []
         for (let i = 0; i < formatTokenList.length; i++) {
           const formatToken = formatTokenList[i]
-          const tokenStringList = formatToken.content.split('')
+          const tokenStringList = splitText(formatToken.content)
           for (let j = 0; j < tokenStringList.length; j++) {
             const value = tokenStringList[j]
             const element: IElement = {